test: extract verified DDO test to separate file, add more checks
