+++ .circleci/config.yml
@@ -670,6 +670,12 @@ workflows:
           suite: itest-direct_data_onboard
           target: "./itests/direct_data_onboard_test.go"
       - test:
+          name: test-itest-direct_data_onboard_verified
+          requires:
+            - build
+          suite: itest-direct_data_onboard_verified
+          target: "./itests/direct_data_onboard_verified_test.go"
+      - test:
           name: test-itest-dup_mpool_messages
           requires:
             - build
